+++ RailwayTicketManager/Passenger_data_handler.h
@@ -8,7 +8,7 @@ struct Passenger {
 	string name;
 	string surname;
 	long tckn;
-	int age;
+	long age;
 };
 
 vector<Passenger> read_passenger_data_from_csv();
@@ -17,4 +17,4 @@ Passenger find_passenger_by_id(long id);
 Passenger find_passenger_by_name(string name, string surname);
 void add_new_passenger();
 long read_last_passenger_id();
-void dump_passenger_csv_data(int num_lines);
+void dump_passenger_csv_data(long num_lines);
