+++ RailwayTicketManager/Ticket_data_handler.cpp
@@ -74,7 +74,7 @@ void save_ticket_data_to_csv(const Ticket& new_ticket) {
 	}
 }
 
-int get_ticket_type(long id) {
+long get_ticket_type(long id) {
 	Ticket ticket = find_ticket_by_id(id);
 	return ticket.type;
 }
@@ -106,7 +106,7 @@ long read_last_ticket_id() {
 }
 
 void add_new_ticket() {
-	int type, assigned_campaign_id, assigned_passenger_id, assigned_seat_id, new_ticket_id;
+	long type, assigned_campaign_id, assigned_passenger_id, assigned_seat_id, new_ticket_id;
 	// Get the train id from user
 	vector<Seat> seat{};
 	assigned_campaign_id = 1;//ui_campaign_selector();
@@ -133,7 +133,7 @@ void dump_ticket_csv_data(long num_lines) {
 	vector<Ticket> tickets = read_ticket_data_from_csv();
 
 	// Print the specified number of lines
-	int lines_printed = 0;
+	long lines_printed = 0;
 	for (Ticket t : tickets) {
 		// Print the ticket information
 		std::cout << "Ticket ID: " << t.id << endl;
